apirestful/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
pytest==8.1.1
httpx==0.27
passlib==1.7.4
bcrypt==4.1.3
pyjwt==2.8.0